% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{add_sample_path}
\alias{add_sample_path}
\title{Add Salmon output paths to sample table}
\usage{
add_sample_path(sampleDir, sample_table)
}
\arguments{
\item{sampleDir}{Character. Path to the directory containing Salmon output folders, one per sample.}

\item{sample_table}{Data frame with sample metadata. One of its columns should match the folder names in \code{sampleDir}.}
}
\value{
A data frame identical to \code{sample_table}, with an additional \code{Folder} column containing full paths.
Missing folders will result in \code{NA} entries.
}
\description{
This function adds a \code{Folder} column to the \code{sample_table} data frame, indicating the
path where Salmon quantification output (e.g. quant.sf) is expected for each sample, based
on folder names found in the \code{sampleDir} path.
}
