% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{add_to_db}
\alias{add_to_db}
\title{Add key/id file_path pair to db.json.}
\usage{
add_to_db(db_fpath, id, fpaths, sort = FALSE)
}
\arguments{
\item{db_fpath}{file path to database - "../inst/extdata/db.json"}

\item{id}{unique key/id for group of samples}

\item{fpaths}{vector of file paths (biological replicates)}

\item{sort}{sort the keys/ids alphabetically?}
}
\value{
nothing
}
\description{
Add key/id file_path pair to db.json.
}
