% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{analyze}
\alias{analyze}
\title{Perform automated analysis.}
\usage{
analyze(
  keys,
  db,
  meta_dir,
  project_name,
  source,
  sep,
  names,
  out_dir,
  alpha = 0.05,
  lFC = 1,
  ks = "12",
  go = "skip",
  goi = NULL
)
}
\arguments{
\item{keys}{list of keys for db. Order determines testing!: First element -> "denom", Second element: "num", rest is ""}

\item{db}{contains key and list of paths (vector of paths) to key files}

\item{meta_dir}{the directory to which meta file should be written to. Name is generated automatically using project_name and source}

\item{project_name}{inserted as signifier into meta filename}

\item{source}{inserted as signifier into meta filename}

\item{sep}{separator in meta file - tab}

\item{names}{names to be used instead of keys}

\item{out_dir}{output directory for the analysis}

\item{alpha}{p-value for DESeq2 analysis}

\item{lFC}{log2FoldChange threshold for significance}

\item{ks}{number of clusters}

\item{go}{"skip" or don't skip "yes"}

\item{goi}{gene of interest for iVolcano and iMDS plot svg}
}
\value{
nothing
}
\description{
Perform automated analysis.
}
