% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topGO_All.R
\name{analyze_GO_similarity}
\alias{analyze_GO_similarity}
\title{Analyze GO term similarity and visualize clusters}
\usage{
analyze_GO_similarity(
  go_results,
  orgdb = "org.At.tair.db",
  ontology = "BP",
  semdata = NULL,
  output_prefix = NULL,
  Number_GOs = 20,
  units = "in",
  width = 10,
  height = 8
)
}
\arguments{
\item{go_results}{A data.frame with GO enrichment results (must include \code{pval}).}

\item{orgdb}{OrgDb package name (e.g., "org.At.tair.db").}

\item{ontology}{GO ontology used ("BP", "MF", or "CC"). Defaults to "BP".}

\item{semdata}{Precomputed semantic data (optional).}

\item{output_prefix}{Optional prefix for saving plots (PDFs).}

\item{Number_GOs}{Number of top GO term names to plot in the scatterplot. Defaults to 20.}

\item{units}{Units to set plot dimensions. Default to "in" (inches)}

\item{width}{Width units for the plot. Default to 10.}

\item{height}{Height units for the plot. Default to max(4, round(8 * nrow(go_results) / 50)), which was empirically obtained to get a nice heigh depending on the number of enriched GO terms.}
}
\value{
A list with the similarity matrix, reduced terms, and ggplot objects.
}
\description{
This function calculates semantic similarity between GO terms, reduces redundancy,
and generates scatterplot and treemap visualizations.
}
