% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{build_two_level_list}
\alias{build_two_level_list}
\title{Group list of paths by two layers.}
\usage{
build_two_level_list(vec)
}
\arguments{
\item{vec}{the vector of paths.}
}
\value{
list of grouped file paths - 2-layer grouping
}
\description{
Group 2-layer list of paths.
}
