% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_analysis.R
\name{calculate_pca}
\alias{calculate_pca}
\title{Calculate PCA from a DESeq2 object}
\usage{
calculate_pca(dds, ntop = 500)
}
\arguments{
\item{dds}{A DESeqDataSet object.}

\item{ntop}{Integer. Number of top variable genes to use. Default is 500.}
}
\value{
A list with PCA results, percent variance, and the transformed object.
}
\description{
This function performs variance-stabilizing transformation (VST),
selects the top variable genes, and computes PCA.
}
