% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{filter_low_counts}
\alias{filter_low_counts}
\title{Filter lowly expressed genes}
\usage{
filter_low_counts(dds, min_count = 10, min_samples = 3)
}
\arguments{
\item{dds}{A DESeqDataSet object.}

\item{min_count}{Minimum count threshold per sample. Default to 10 TPM.}

\item{min_samples}{Minimum number of samples that must meet the threshold. Deafult to 3.}
}
\value{
A filtered DESeqDataSet object.
}
\description{
This function removes genes with low expression across samples, which may introduce noise in downstream analyses.
A gene is retained if it has at least \code{min_count} TPMs in at least \code{min_samples} samples of the analysis.
}
