% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topGO_All.R
\name{GeneNames_GOs}
\alias{GeneNames_GOs}
\title{GeneNames_GOs: Generate Excel report of genes associated with GO terms}
\usage{
GeneNames_GOs(Annotation, geneID2GO, output_dir, DEG, Ontologies = NULL, name)
}
\arguments{
\item{Annotation}{Two column data frame. First column must include Gene IDs.
Second column must include functional annotation.}

\item{geneID2GO}{Named list mapping gene IDs to GO terms.}

\item{output_dir}{Directory where the output file will be saved.}

\item{DEG}{Vector or Data frame. If vector, it must include Gene IDs of interest.
If Data frame, it must be the output of differential expressed genes (row names should be Gene IDs).}

\item{Ontologies}{Character vector of GO terms to search for.}

\item{name}{Base name for the output file.}
}
\value{
No return value. An Excel file is written to disk.
}
\description{
This function generates an Excel report of genes associated with specified GO terms.
It reads a functional annotation file, filters genes based on the provided DEGs and GO terms,
and saves the results in an Excel file.
}
