% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_sample_subset}
\alias{get_sample_subset}
\title{Subset the sample table based on selection criteria}
\usage{
get_sample_subset(sample_table, Include = NULL, Exclude = NULL)
}
\arguments{
\item{sample_table}{A data frame containing sample metadata.}

\item{Include}{A vector of condition(s) to include samples from \code{sample_table}.
It will include samples matching this/these condition(s) and exclude all others.
Default to NULL.}

\item{Exclude}{A vector of condition(s) to exclude samples from \code{sample_table}.
Default to NULL.}
}
\value{
A filtered data frame with selected samples.
}
\description{
This function let you select which samples you want to include in the analysis
from those listed in your sample table. Think twice if your analysis requires a selection
or not. Use the \code{selection} argument  carefully, as it can be misleading. Always verify that the correct samples
have been selected before proceeding with downstream analysis.
}
