% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_all.R
\name{load_module_counts}
\alias{load_module_counts}
\title{Load expression data for a module}
\usage{
load_module_counts(color, TPM_filt, geneInfo)
}
\arguments{
\item{color}{Module color.}

\item{TPM_filt}{Filtered TPM expression matrix.}

\item{geneInfo}{Data frame with gene-module assignments.}
}
\value{
A data frame with expression values for genes in the module.
}
\description{
Load expression data for a module
}
