% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_topGO_db.R
\name{load_topGO_db}
\alias{load_topGO_db}
\title{Creation of a correct database for topGO}
\usage{
load_topGO_db(Annotation_file, GeneID, GOterms)
}
\arguments{
\item{Annotation_file}{Data frame including the annotation data of your organism. This must contain at least one column with the GeneID and another with GO term(s)}

\item{GeneID}{Name of the column with GeneIDs}

\item{GOterms}{Name of the column with GOterms representing GeneIDs}
}
\value{
List in the appropiate topGO format
}
\description{
This function adapts your annotation file from your organism of interest to the proper formats that topGO requieres for GO enrichment analysis
}
\examples{
# Read data with DEGs from DESeq2
DEGs = read.table()
# Perform GO enrichment of BP with Tak-1 OPDA vs Tak-1 Mock differentially expressed genes
read_table()

}
