% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_tximport_data}
\alias{load_tximport_data}
\title{Load expression data using tximport}
\usage{
load_tximport_data(sampleDir, sample_table, tx2gene)
}
\arguments{
\item{sampleDir}{Directory containing Salmon quantification files.}

\item{sample_table}{A data frame with sample metadata.}

\item{tx2gene}{A data frame with columns TXNAME and GENEID (transcript-to-gene mapping).}
}
\value{
A tximport object with abundance data.
}
\description{
This function loads transcript quantification data generated by Salmon using \code{tximport}.
It requires a \code{tx2gene} data frame to map transcripts to genes.
}
