% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{match_beginning}
\alias{match_beginning}
\title{Group list by vector containing all beginning strings.}
\usage{
match_beginning(vec, kinds)
}
\arguments{
\item{vec}{the vector of paths.}

\item{kinds}{the vector of beginning strings which becomes name of the result list}
}
\value{
list of grouped file paths
}
\description{
Group list of paths by filename beginnings. It sorts by longest forms first and
removes all elements which were already chosen previously in the list.
}
