% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_all.R
\name{normalize_df}
\alias{normalize_df}
\title{Normalize expression matrix by gene}
\usage{
normalize_df(df)
}
\arguments{
\item{df}{A numeric matrix or data frame.}
}
\value{
A normalized matrix.
}
\description{
This function scales each gene's expression values between 0 and 1.
}
