% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_simple.R
\name{plot_deseq_heatmap}
\alias{plot_deseq_heatmap}
\title{Plot heatmap of significant DEGs}
\usage{
plot_deseq_heatmap(
  dds,
  res,
  variables,
  name,
  output_dir,
  width = 6,
  height = 8
)
}
\arguments{
\item{dds}{A DESeqDataSet object.}

\item{res}{A DESeqResults object.}

\item{variables}{Character vector of column names in \code{sample_table} to annotate samples in heatmap.}

\item{name}{Name used for saving the plot.}

\item{output_dir}{Directory where the plot will be saved.}

\item{width}{Width inches for the plot. Default to 6.}

\item{height}{Height inches for the plot. Default to 8.}
}
\value{
No return value. A PDF heatmap is saved.
}
\description{
This function generates a heatmap of genes with adjusted p-value <= 0.05 using normalized counts.
}
