% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_all.R
\name{plot_module_trait_relationships}
\alias{plot_module_trait_relationships}
\title{Plot module-trait relationships}
\usage{
plot_module_trait_relationships(net, datExpr, datTraits, output_path)
}
\arguments{
\item{net}{WGCNA network object.}

\item{datExpr}{Expression matrix (samples x genes).}

\item{datTraits}{Trait matrix (samples x traits).}

\item{output_path}{Directory where the plot will be saved.}
}
\value{
No return value. A PDF file is saved.
}
\description{
This function generates a heatmap of correlations between module eigengenes and traits.
}
