% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_analysis.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA from PCA results}
\usage{
plot_pca(
  pca_result,
  variables,
  colors,
  components = c(1, 2),
  name = "PCA",
  output_dir = ".",
  width = 10,
  height = 8,
  units = "in"
)
}
\arguments{
\item{pca_result}{Output from \code{calculate_pca()}.}

\item{variables}{Character vector of two column names from \code{sample_table} used for grouping.
The first variable will determine the point shapes, and the second will determine the point colors in the PCA plot.}

\item{colors}{Named vector of colors for groups.}

\item{components}{Integer vector of length 2 indicating which PCs (1 to 4) to plot}

\item{name}{Character. Name used for saving the plot.}

\item{output_dir}{Directory where the plot will be saved.}

\item{width}{Width units for the plot. Default to 10.}

\item{height}{Height units for the plot. Default to 8.}

\item{units}{Units to set plot dimensions. Default to "in" (inches)}
}
\value{
A ggplot object.
}
\description{
This function generates a PCA plot using ggplot2.
}
