% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_all.R
\name{plot_sample_clustering}
\alias{plot_sample_clustering}
\title{Plot sample clustering dendrogram}
\usage{
plot_sample_clustering(datExpr, output_path)
}
\arguments{
\item{datExpr}{Expression matrix (samples x genes).}

\item{output_path}{Directory where the plot will be saved.}
}
\value{
No return value. A PDF file is saved.
}
\description{
This function generates a hierarchical clustering dendrogram of samples.
}
