% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_all.R
\name{plot_soft_threshold}
\alias{plot_soft_threshold}
\title{Plot soft-thresholding power selection}
\usage{
plot_soft_threshold(sft, output_path)
}
\arguments{
\item{sft}{Output from WGCNA::pickSoftThreshold().}

\item{output_path}{Directory where the plot will be saved.}
}
\description{
This function plots the scale-free topology fit index for various soft-thresholding powers.
}
