% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topGO_All.R
\name{prepare_topGO_genes}
\alias{prepare_topGO_genes}
\title{Prepare genes for topGO analysis}
\usage{
prepare_topGO_genes(DEG, geneUniverse, padj_threshold = 0.05)
}
\arguments{
\item{DEG}{Vector of genes or data.frame with differential expression results. Genes must be rownames in the data.frame.}

\item{geneUniverse}{Vector with all gene names of the organism universe (e.g., \code{geneID2GO} names).}

\item{padj_threshold}{Significance threshold for \code{padj} if \code{DEG} is a data.frame.}
}
\value{
Binary factor with gene names, to be used in \code{topGO}.
}
\description{
This function takes a gene vector or a data.frame with differential expression results
and returns a binary factor that shows which genes are in the list of interest.
}
