% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{remove_from_db}
\alias{remove_from_db}
\title{Remove entry (key/id) from db.json.}
\usage{
remove_from_db(id, db_fpath = NULL)
}
\arguments{
\item{id}{unique key/id for group of samples}

\item{db_fpath}{file path to database - default is "../inst/extdata/db.json"}
}
\value{
nothing
}
\description{
Remove entry (key/id) from db.json.
}
