% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_analysis.R
\name{run_pca_analysis}
\alias{run_pca_analysis}
\title{Run full PCA analysis and save plot}
\usage{
run_pca_analysis(
  dds,
  variables,
  colors,
  components = c(1, 2),
  name = "PCA",
  output_dir = ".",
  ntop = 500,
  width = 10,
  height = 8,
  units = "in"
)
}
\arguments{
\item{dds}{A DESeqDataSet object.}

\item{variables}{Character vector of two column names from \code{sample_table} used for grouping.
The first variable will determine the point shapes, and the second will determine the point colors in the PCA plot.}

\item{colors}{Named vector of colors for groups.}

\item{components}{Integer vector of length 2 indicating which PCs (1 to 4) to plot}

\item{name}{Character. Name used for saving the plot.}

\item{output_dir}{Directory where the plot will be saved.}

\item{ntop}{Integer. Number of top variable genes to use. Default is 500.}

\item{width}{Width units for the plot. Default to 10.}

\item{height}{Height units for the plot. Default to 8.}

\item{units}{Units to set plot dimensions. Default to "in" (inches)}
}
\value{
A ggplot object.
}
\description{
This function runs the full PCA pipeline: transformation, PCA computation, and plotting.
}
