% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_all.R
\name{save_module_info}
\alias{save_module_info}
\title{Save module information to CSV}
\usage{
save_module_info(net, datExpr, datTraits, output_path, Name)
}
\arguments{
\item{net}{WGCNA network object.}

\item{datExpr}{Expression matrix (samples x genes).}

\item{datTraits}{Trait matrix (samples x traits).}

\item{output_path}{Directory where the output file will be saved.}

\item{Name}{Character. Base name for output files.}
}
\value{
No return value. A CSV file is written.
}
\description{
This function saves the gene-module assignments and calculates module-trait correlations.
}
