% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topGO_All.R
\name{save_topGO_results}
\alias{save_topGO_results}
\title{Save topGO results to a text file}
\usage{
save_topGO_results(go_results, output_path, simplified = FALSE)
}
\arguments{
\item{go_results}{Data.frame with GO results.}

\item{output_path}{Path to save the file.}

\item{simplified}{Logical. If TRUE, only p-values are saved.}
}
\description{
Save topGO results to a text file
}
