% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\docType{data}
\name{sym2len}
\alias{sym2len}
\title{Gene lengths from gtf mm10.}
\format{
numeric named vector: numbers are gene length name is symbol.
}
\source{
\url{}
}
\usage{
sym2len
}
\description{
symbol to gene length gtf mm10.
}
\keyword{datasets}
