% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPM_all.R
\name{TPM_all}
\alias{TPM_all}
\title{Join TPM values from all samples in a data frame}
\usage{
TPM_all(
  sampleDir,
  sample_table,
  output_path,
  Include = NULL,
  Exclude = NULL,
  Save_results = TRUE,
  tx2gene
)
}
\arguments{
\item{sampleDir}{Directory containing sample folders with quant.sf files.}

\item{sample_table}{Data frame with sample metadata.}

\item{output_path}{Directory where final data frame will be saved.}

\item{Include}{A vector of condition(s) to include samples from \code{sample_table}.
It will include samples matching this/these condition(s) and exclude all others.
Default to NULL.}

\item{Exclude}{A vector of condition(s) to exclude samples from \code{sample_table}.
Default to NULL.}

\item{Save_results}{Logical. Whether to save the final data frame in a .txt file. Default is TRUE}

\item{tx2gene}{A data frame with columns TXNAME and GENEID (transcript-to-gene mapping).}
}
\value{
No return value. Files are written to disk.
}
\description{
This function reads TPM quantified data from Salmon and joins them into a final data frame.
Results are optionally saved in a .txt file
}
