% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_all.R
\name{WGCNA_Modules}
\alias{WGCNA_Modules}
\title{WGCNA Modules Analysis}
\usage{
WGCNA_Modules(
  output_path,
  sampleDir,
  sample_table,
  Include = NULL,
  Exclude = NULL,
  DEGs,
  Variables,
  tx2gene,
  Filter = NULL,
  Power,
  Name,
  Colors_plot = NULL,
  NumberCol = 1,
  geneID2GO,
  Annotation,
  ...
)
}
\arguments{
\item{output_path}{Directory where results will be saved.}

\item{sampleDir}{Directory containing sample folders with quant.sf files.}

\item{sample_table}{Data frame with sample metadata.}

\item{DEGs}{Vector of differentially expressed genes.}

\item{Variables}{Character vector of column names in sample_table used for grouping.}

\item{tx2gene}{A data frame with columns TXNAME and GENEID (transcript-to-gene mapping).}

\item{Filter}{Character vector of samples to exclude.}

\item{Power}{Soft-thresholding power for WGCNA.}

\item{Name}{Character. Base name for output files.}

\item{Colors_plot}{Named vector of colors for groups.}

\item{NumberCol}{Integer. Number of columns for facet_wrap in ggplot.}

\item{geneID2GO}{Named list mapping gene IDs to GO terms.}

\item{Annotation}{Two column data frame. First column must include Gene IDs.
Second column must include functional annotation.}

\item{Selection}{Named vector indicating samples to include/exclude.}
}
\value{
A list with WGCNA results and plots.
}
\description{
This function performs WGCNA analysis to identify modules of co-expressed genes and their association with traits.
}
