% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topGO_All.R
\name{plot_topGO_results}
\alias{plot_topGO_results}
\title{Plot topGO enrichment results}
\usage{
plot_topGO_results(
  go_results,
  title = "GO Enrichment",
  output_path = NULL,
  units = "in",
  width = 10,
  height = max(4, round(8 * nrow(go_results)/50))
)
}
\arguments{
\item{go_results}{A data.frame with GO enrichment results (output of \code{process_topGO_results()}).}

\item{title}{Title of the plot.}

\item{output_path}{Optional path to save the plot as PDF. If NULL, the plot is not saved.}

\item{units}{Units to set plot dimensions. Default to "in" (inches)}

\item{width}{Width units for the plot. Default to 10.}

\item{height}{Height units for the plot. Default to max(4, round(8 * nrow(go_results) / 50)), which was empirically obtained to get a nice heigh depending on the number of enriched GO terms.}
}
\value{
A ggplot object.
}
\description{
This function creates a bubble plot of enriched GO terms using ggplot2.
}
