% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_all.R
\name{run_topGO_for_modules}
\alias{run_topGO_for_modules}
\title{Run topGO enrichment analysis for each WGCNA module}
\usage{
run_topGO_for_modules(
  geneID2GO,
  Name,
  input_path,
  output_path_topGO,
  Annotation,
  ...
)
}
\arguments{
\item{geneID2GO}{Named list mapping gene IDs to GO terms.}

\item{Name}{Character. Base name for '"Name"_geneInfo.csv' previously created with WGCNA.}

\item{input_path}{Directory where '"Name"_geneInfo.csv' file was saved.}

\item{output_path_topGO}{Directory where results will be saved.}

\item{Annotation}{Two column data frame. First column must include Gene IDs.
Second column must include functional annotation.}

\item{Ontologies}{Character vector of GO terms of interest to search for.}
}
\value{
A named list with GO enrichment results per module.
}
\description{
This function performs GO enrichment analysis using topGO for each module identified by WGCNA.
}
